Short circuit combined reducers if none of the reducers respond to an action
In multiply nested `combineReducer` trees, use available information at
the time the reducer is called to construct a total set of all actions
responded to by reducers contained in the combined reducer. If an action
is passed to the reducer which isn't used by any contained reducer, and
the state is already defined, short circuit and return the state instead
of iterating through each key and calling the reducer.